requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
better-exceptions==0.2.2
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.8.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
greenlet==0.4.15
grpcio==1.18.0
gym==0.8.2
h5py==2.9.0
idna==2.8
imageio==2.4.1
imageio-ffmpeg==0.2.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
moviepy==0.2.3.5
mpi4py==2.0.0
msgpack==0.6.1
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pybullet==1.9.2
pyglet==1.3.2
Pygments==2.3.1
pynvim==0.3.2
PyOpenGL==3.1.0
pyparsing==2.3.1
python-dateutil==2.8.0
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.21.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
